+++ webview-ui/src/app/app.component.html
@@ -1,12 +1,4 @@
-<!-- app.component.html -->
 <div>
-  <label for="filePathInput">Enter File Path:</label>
-  <input type="text" id="filePathInput" [(ngModel)]="filePath" />
-  <button (click)="loadNetwork()">Load Network</button>
+  <app-folder-file></app-folder-file>
 </div>
-
-<div
-  #networkContainer
-  style="width: 2000px; height: 1500px; background-color: white"
-></div>
 <router-outlet></router-outlet>
