+++ webview-ui/src/app/app.module.ts
@@ -5,9 +5,10 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 
 import { FormsModule } from '@angular/forms';
+import { FolderFileComponent } from './folder-file/folder-file.component';
 
 @NgModule({
-  declarations: [AppComponent],
+  declarations: [AppComponent, FolderFileComponent],
   imports: [BrowserModule, AppRoutingModule, FormsModule],
   providers: [],
   bootstrap: [AppComponent],
