+++ webview-ui/src/app/folder-file/folder-file.component.html
@@ -1,0 +1,12 @@
+<!-- app.component.html -->
+<div>
+  <label for="filePathInput">Enter File Path:</label>
+  <input type="text" id="filePathInput" [(ngModel)]="filePath" />
+  <button (click)="loadNetwork()">Load Network</button>
+</div>
+
+<div
+  #networkContainer
+  style="width: 2000px; height: 1500px; background-color: white"
+></div>
+<router-outlet></router-outlet>
